KG-146. Allow feature to disable filtering for events
- OpenTelemetry and Debugger features relay on the agent execution flow and the sequence of events to produce correct spans. Adding filter for the feature might break the feature logic. Make it possible to disable the ability to filter events on feature level;
- Disable events filtering for OpenTelemetry and Debugger features.
