+++ agents/agents-core/src/commonMain/kotlin/ai/koog/agents/core/feature/config/FeatureConfig.kt
@@ -52,7 +52,7 @@ public abstract class FeatureConfig {
      * }
      * ```
      */
-    public fun setEventFilter(filter: (AgentLifecycleEventContext) -> Boolean) {
+    public open fun setEventFilter(filter: (AgentLifecycleEventContext) -> Boolean) {
         _eventFilter = filter
     }
 }
