+++ agents/agents-core/src/commonMain/kotlin/ai/koog/agents/core/feature/debugger/DebuggerConfig.kt
@@ -1,6 +1,7 @@
 package ai.koog.agents.core.feature.debugger
 
 import ai.koog.agents.core.feature.config.FeatureConfig
+import ai.koog.agents.core.feature.handler.AgentLifecycleEventContext
 import kotlin.time.Duration
 
 /**
@@ -45,5 +46,11 @@ public class DebuggerConfig : FeatureConfig() {
      */
     public fun setAwaitInitialConnectionTimeout(timeout: Duration) {
         _awaitInitialConnectionTimeout = timeout
+    }
+
+    override fun setEventFilter(filter: (AgentLifecycleEventContext) -> Boolean) {
+        // Do not allow events filtering for the Debugger feature.
+        // Debugger relays on the execution sequence. Filtering events can break the feature logic.
+        throw UnsupportedOperationException("Events filtering is not allowed for the Debugger feature.")
     }
 }
